[lexical-react][lexical-table] Remove hook and move logic to lexical-table package
