--- packages/lexical-react/flow/useLexicalTableObservers.js.flow
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow strict
- */
-
-import type {LexicalEditor, NodeKey} from 'lexical';
-
-declare export function useLexicalTableObservers(
-  editor: LexicalEditor,
-  hasTabHandler: boolean,
-): { refreshTableObservers: (nodeKey: NodeKey) => void; };
