+++ packages/lexical-table/src/index.ts
@@ -30,8 +30,10 @@ export {
 export type {TableDOMCell} from './LexicalTableObserver';
 export {$getTableAndElementByKey, TableObserver} from './LexicalTableObserver';
 export {
+  registerOptimizedTableSelectionObserver,
   registerTableCellUnmergeTransform,
   registerTablePlugin,
+  registerTableSelectionObserver,
 } from './LexicalTablePluginHelpers';
 export type {SerializedTableRowNode} from './LexicalTableRowNode';
 export {
@@ -54,27 +56,10 @@ export type {HTMLTableElementWithWithTableSelectionState} from './LexicalTableSe
 export {
   $addHighlightStyleToTable,
   $findCellNode,
-  $findParentTableCellNodeInTable,
   $findTableNode,
-  $getObserverCellFromCellNodeOrThrow,
-  $getTableEdgeCursorPosition,
-  $handleArrowKey,
-  $insertParagraphAtTableEdge,
-  $isFullTableSelection,
-  $isSelectionInTable,
-  $removeHighlightStyleToTable,
-  $selectAdjacentCell,
-  ARROW_KEY_COMMANDS_WITH_DIRECTION,
-  DELETE_KEY_COMMANDS,
-  DELETE_TEXT_COMMANDS,
   getDOMCellFromTarget,
-  getDOMCellInTableFromTarget,
-  getEditorWindow,
   getTableElement,
   getTableObserverFromTableElement,
-  isPointerDownOnEvent,
-  LEXICAL_ELEMENT_KEY,
-  stopEvent,
 } from './LexicalTableSelectionHelpers';
 export {
   $computeTableMap,
